requirements.txt
streamlit
openai
scikit-learn
python-dotenv
pytest